requirements.txt
p2pnetwork == 1.1
pyfiglet == 0.8.post1
pycryptodome == 3.14.1